Hpcl code new

+++ UrdhvaBase/urdhva_base/postgresmodel.py
@@ -1,3 +1,4 @@
+import re
 import json
 import time
 import typing
@@ -5,6 +6,7 @@ import asyncio
 import pydantic
 import datetime
 import traceback
+import collections
 import urdhva_base
 import urdhva_base.settings
 import urdhva_base.redispool
@@ -106,6 +108,68 @@ class BasePostgresModel(pydantic.BaseModel):
         ...
 
     @classmethod
+    async def get_clause_conditions(cls, formated=False, extra_key_mapping=None, default_mapping=None):
+        where_clause = []
+        if urdhva_base.ctx.exists() and hasattr(cls.Config, "access_key_mapping"):
+            key_mapping = cls.Config.access_key_mapping
+            mapped_data = {
+                key.split(":")[0].strip(): key.split(":")[1].strip() if ":" in key else key.split(":")[0].strip()
+                for key in key_mapping
+            }
+            # Generating OR Conditions for mapped keys
+            or_condition_keys = {}
+            for key in [item for item, count in collections.Counter(list(mapped_data.values())).items() if count > 1]:
+                for key_, value in mapped_data.items():
+                    if value == key:
+                        if key not in or_condition_keys:
+                            or_condition_keys[key] = []
+                        or_condition_keys[key].append(key_)
+            # Cleaning if key matching or_conditions and not equals to base key
+            for base_key, mapped_keys in or_condition_keys.items():
+                for mapped_key in mapped_keys:
+                    if mapped_key != base_key and mapped_key in mapped_data:
+                        del mapped_data[mapped_key]
+            rpt = urdhva_base.context.context.get('rpt', {})
+            # Removing BU incase if SAP_ID was available and restricted
+            # Todo:- Need to verify in multiple conditions
+            if 'bu' in mapped_data and 'sap_id' in mapped_data:
+                if rpt.get('bu') and rpt.get('sap_id'):
+                    del mapped_data['bu']
+            # Generating query Conditions
+            for key, value in mapped_data.items():
+                if extra_key_mapping and key in extra_key_mapping:
+                    key = extra_key_mapping[key]
+                # Incase if we need to map extra details like bu for supply chain
+                if default_mapping and key in default_mapping:
+                    value = default_mapping[key]
+                if value in rpt and rpt.get(value):
+                    if isinstance(rpt[value], list):
+                        if len(rpt[value]) == 1:
+                            if formated:
+                                where_clause.append({'key': key, "cond": 'equals', "value": rpt[value][0]})
+                            else:
+                                if key in or_condition_keys:
+                                    conditions = [f"{k}='{rpt[value][0]}'" for k in or_condition_keys[key]]
+                                    where_clause.append(f"({' OR '.join(conditions)})")
+                                else:
+                                    where_clause.append(f"{key}='{rpt[value][0]}'")
+                        else:
+                            if formated:
+                                where_clause.append({'key': key, "cond": ' ', "value": rpt[value]})
+                            else:
+                                if key in or_condition_keys:
+                                    conditions = [f"{k} in {tuple(rpt[value])}" for k in or_condition_keys[key]]
+                                    where_clause.append(f"({' OR '.join(conditions)})")
+                                else:
+                                    where_clause.append(f"{key} in {tuple(rpt[value])}")
+                    else:
+                        if formated:
+                            where_clause.append({'key': key, "cond": 'equals', "value": rpt[value]})
+                        else:
+                            where_clause.append(f"{key}='{rpt[value]}'")
+        return where_clause
+
+    @classmethod
     async def _get_entity_id(cls, entity_id=None):
         if urdhva_base.ctx.exists():
             return urdhva_base.ctx['entity_id']
@@ -131,7 +195,7 @@ class BasePostgresModel(pydantic.BaseModel):
         return resp
 
     @classmethod
-    async def count(cls, params: urdhva_base.queryparams.QueryParams = None, entity_id=None):
+    async def count(cls, params: urdhva_base.queryparams.QueryParams = None, entity_id=None, count_query=None):
         """
 
         :param params:
@@ -142,34 +206,51 @@ class BasePostgresModel(pydantic.BaseModel):
         # if not entity_id:
         #     raise "missing context information"
         query = [f"select count('*') from {cls.__tablename__}"]
-        query_params = [f"{cls.__tablename__}.entity_id = '{entity_id}'"] if entity_id else []
-        # todo:- need to implement this for all keys, both permitted and prohibited
-        if urdhva_base.ctx.exists() and urdhva_base.context.context.get('rpt', {}):
-            rpt = urdhva_base.context.context.get('rpt', {})
-            key = f"access_restrictions_{urdhva_base.ctx['entity_id']}"
-            redis_client = await urdhva_base.redispool.get_redis_connection()
-            if await redis_client.hexists(key, rpt['email']):
-                data = json.loads(await redis_client.hget(f"access_restrictions_{urdhva_base.ctx['entity_id']}",
-                                                          rpt['email']))
-                # todo:- Temporary fix for ACL'S
-                if data.get("organizations_permitted"):
-                    permitted_org = data['organizations_permitted'].split(",")
-                    if len(permitted_org) == 1:
-                        permitted_org = f"('{permitted_org[0]}')"
-                    if cls.__tablename__ == "organization":
-                        query_params.append(f"id in {permitted_org}")
-                    else:
-                        query_params.append(f"organization_id in {permitted_org}")
-        if hasattr(cls.Config, 'standard_query'):
-            query_params.append(cls.Config.standard_query)
-        if params and params.q and len(params.q) > 0:
-            query_params.append(params.q)
-        if len(query_params):
-            query.append("where " + " AND ".join(query_params))
+        if not count_query or not isinstance(count_query, list):
+            query_params = [f"{cls.__tablename__}.entity_id = '{entity_id}'"] if entity_id else []
+            # todo:- need to implement this for all keys, both permitted and prohibited
+            if urdhva_base.ctx.exists() and urdhva_base.context.context.get('rpt', {}):
+                rpt = urdhva_base.context.context.get('rpt', {})
+                key = f"access_restrictions_{urdhva_base.ctx['entity_id']}"
+                redis_client = await urdhva_base.redispool.get_redis_connection()
+                if await redis_client.hexists(key, rpt['email']):
+                    data = json.loads(await redis_client.hget(f"access_restrictions_{urdhva_base.ctx['entity_id']}",
+                                                              rpt['email']))
+                    # todo:- Temporary fix for ACL'S
+                    if data.get("organizations_permitted"):
+                        permitted_org = data['organizations_permitted'].split(",")
+                        if len(permitted_org) == 1:
+                            permitted_org = f"('{permitted_org[0]}')"
+                        if cls.__tablename__ == "organization":
+                            query_params.append(f"id in {permitted_org}")
+                        else:
+                            query_params.append(f"organization_id in {permitted_org}")
+            if hasattr(cls.Config, 'standard_query'):
+                query_params.append(cls.Config.standard_query)
+            query_params.extend(await cls.get_clause_conditions())
+            if params and params.q and len(params.q) > 0:
+                query_params.append(params.q)
+            if len(query_params):
+                query.append("where " + " AND ".join(query_params))
+        else:
+            query.extend(count_query[1:])
         session = await manager.get_session()
         total = await session.scalar(text(" ".join(query)))
         await asyncio.shield(session.close())
         return total
+
+    @classmethod
+    async def update_by_query(cls, query, entity_id=None):
+        session = await manager.get_session()
+        try:
+            result = await session.execute(text(query))
+            print(f"Rows committed {result.rowcount}")
+            await session.commit()
+        except Exception as e:
+            print(f"Exception while running update by query {e}")
+            raise f"Exception while running update by query {e}"
+        finally:
+            await asyncio.shield(session.close())
 
     @classmethod
     async def get_aggr_data(cls, query, limit=100, skip=0, skip_total=True):
@@ -191,7 +272,7 @@ class BasePostgresModel(pydantic.BaseModel):
                 query_ = f"{query} LIMIT {limit} OFFSET {limit * skip}"
             else:
                 query_ = f"{query}"
-            if not query_.upper().startswith("WITH ") and not query_.upper().startswith("SELECT "):
+            if not query_.strip().upper().startswith("WITH ") and not query_.strip().upper().startswith("SELECT "):
                 query_ = f"select {query_}"
             result = await session.execute(text(query_))
             resp = result.all()
@@ -267,15 +348,50 @@ class BasePostgresModel(pydantic.BaseModel):
             query_params.append(cls.Config.standard_query)
         if params and params.q and len(params.q) > 0:
             query_params.append(params.q)
+        # if params and params.q:
+        #     search_conditions = []
+        #     if hasattr(cls.Config, 'search_fields') and cls.Config.search_fields:
+        #         for field in cls.Config.search_fields:
+        #             if params.search_text:
+        #                 search_conditions.append(f"{cls.__tablename__}.{field} ILIKE '%{params.search_text}%'")
+                
+        #         if search_conditions:
+        #             print("search_conditions --> ", search_conditions)
+        #             query_params.append(f"({' OR '.join(search_conditions)})")
+        
+        if params and params.search_text and isinstance(params.search_text, str) and params.search_text.strip():
+            search_conditions = []
+            if hasattr(cls.Config, 'search_fields') and cls.Config.search_fields:
+                for field in cls.Config.search_fields:
+                    search_text = params.search_text.strip()  # Strip leading/trailing whitespace
+                    print("search text --> ", search_text)
+                    search_conditions.append(f"{cls.__tablename__}.{field}::text ILIKE '%{search_text}%'")
+
+            if search_conditions:
+                print("search_conditions --> ", search_conditions)
+                query_params.append(f"({' OR '.join(search_conditions)})")        
+        
+        # Commented by Shrihari
+        # query_params.extend(await cls.get_clause_conditions())
+        # Added to remove the duplicate condition of bu
+        _extra_condition = await cls.get_clause_conditions()
+        pattern = r"\s*bu\s*=\s*'[^']+'\s*(\s*)?"
+        cleaned_conditions = []
+        for cond in _extra_condition:
+            new_cond = re.sub(pattern, '', cond).strip()
+            if new_cond:
+                cleaned_conditions.append(new_cond.rstrip("AND").rstrip("and"))
+        query_params.extend(cleaned_conditions)
         if len(query_params):
             query.append("where " + " AND ".join(query_params))
+        count_query = [rec for rec in query]
 
         # ******************** For Data Sorting Params Start ******************** #
         if params.sort:
             try:
                 order_by = json.loads(params.sort) if isinstance(params.sort, str) else params.sort
                 for key, value in order_by.items():
-                    order = 'ASC' if 'asc' in value.lower() else 'DESC'
+                    order = 'ASC' if 'asc' in value.lower() else 'DESC'  
                     query.append(f"ORDER BY {key} {order}")
                     break
             except Exception as e:
@@ -293,7 +409,7 @@ class BasePostgresModel(pydantic.BaseModel):
             result = await session.scalars(select(cls.Config.schema_class).from_statement(text(" ".join(query))))
             resp = result.all()
             results = [{key: value for key, value in row.__dict__.items() if not key.startswith("_")} for row in resp]
-            total = await cls.count(params, entity_id)
+            total = await cls.count(params, entity_id, count_query) if len(results) > 0 else 0
             results_data = {"data": results, "count": len(results), "total": total}
             if resp_type == "encoded":
                 return JSONResponse(content=jsonable_encoder(results_data))
@@ -321,14 +437,21 @@ class BasePostgresModel(pydantic.BaseModel):
         # if not entity_id:
         #     raise "missing context information"
         session = await manager.get_session()
-        await session.execute(
-            cls.Config.schema_class.__table__.delete().where(cls.Config.schema_class.id == int(row_id))
+        result = await session.execute(
+            select(cls.Config.schema_class).where(cls.Config.schema_class.id == int(row_id))
         )
-        await session.commit()
-        await asyncio.shield(session.close())
-        return True, "Success"
 
-    async def create(self, entity_id=None, upsert=False, upsert_skip_keys = []):
+        if result.scalars().first() is not None:
+            status = await session.execute(
+                cls.Config.schema_class.__table__.delete().where(cls.Config.schema_class.id == int(row_id))
+            )
+            
+            await session.commit()
+            await asyncio.shield(session.close())
+            return True, "Success"
+        return False, "Fail"
+
+    async def create(self, entity_id=None, upsert=False, upsert_skip_keys=[]):
         """
 
         :param entity_id:
@@ -343,7 +466,7 @@ class BasePostgresModel(pydantic.BaseModel):
         else:
             upsert_skip_keys = list(set(upsert_skip_keys + ["id", "entity_id", "created_at", "updated_at"]))
 
-        await manager.create_all()
+        # await manager.create_all()
         session = await manager.get_session()
         try:
             if not upsert:
@@ -351,7 +474,7 @@ class BasePostgresModel(pydantic.BaseModel):
                 session.add(schema_class)
                 await session.commit()
                 await session.refresh(schema_class)
-                return schema_class
+                return {"id": schema_class.id, **{key: value for key, value in schema_class.__dict__.items() if not key.startswith("_")}}
             else:
                 ins_resp = insert(self.Config.schema_class).values([{**json.loads(self.model_dump_json()),
                                                                      "entity_id": entity_id}])
@@ -359,7 +482,9 @@ class BasePostgresModel(pydantic.BaseModel):
                 ins_resp = ins_resp.on_conflict_do_update(
                     index_elements=self.Config.upsert_keys, set_=conflict_dict
                 )
-                await session.execute(ins_resp)
+                resp = await session.execute(ins_resp)
+                id = resp.scalar()
+                return {"id": id, **{key: value for key, value in resp.__dict__.items() if not key.startswith("_")}}
         except Exception as e:
             print(f"Exception in {'create' if not upsert else 'upsert'} {e}")
             return None
@@ -384,6 +509,8 @@ class BasePostgresModel(pydantic.BaseModel):
         record = result.one()
         if len(record):
             for key, value in self.model_dump(exclude_none=True, exclude_unset=True).items():
+                if key == 'updated_at':
+                    continue
                 setattr(record[0], key, value)
             await session.commit()
             await asyncio.shield(session.close())
@@ -409,7 +536,7 @@ class BasePostgresModel(pydantic.BaseModel):
         else:
             upsert_skip_keys = list(set(upsert_skip_keys + ["id", "entity_id", "created_at", "updated_at"]))
 
-        await manager.create_all()
+        # await manager.create_all()
         session = await manager.get_session()
         try:
             if not upsert:
@@ -463,6 +590,7 @@ class BasePostgresModel(pydantic.BaseModel):
         from_attributes = True
         collection_name: urdhva_base.settings.default_index
         schema_class: Base
+        search_fields: []
         upsert_keys: []
